requirements.txt
pandas
matplotlib
radon
pydriller
gitpython
lizard
ijson